Move card components into their own file
